Updated jenkins version to 2.49 (alpine) Updated plugins

+++ jenkins-restore-backup.sh
@@ -9,7 +9,11 @@ fi
 if [ -d "$JENKINS_HOME_BACKUP_DIR" ] && [ "$(ls -A $JENKINS_HOME_BACKUP_DIR)" ]; then
   if [ -z "$(ls -lsA --ignore='.*' $JENKINS_HOME | grep -v 'total 0')" ]; then
     JENKINS_LAST_BACKUP=$(ls -1v $JENKINS_HOME_BACKUP_DIR | tail -1)
-    cp -r ${JENKINS_HOME_BACKUP_DIR}/${JENKINS_LAST_BACKUP}/* $JENKINS_HOME
+    if [ -f "${JENKINS_HOME_BACKUP_DIR}/${JENKINS_LAST_BACKUP}" ]; then
+        tar xzfv ${JENKINS_HOME_BACKUP_DIR}/${JENKINS_LAST_BACKUP} $JENKINS_HOME
+    else
+        cp -r ${JENKINS_HOME_BACKUP_DIR}/${JENKINS_LAST_BACKUP}/* $JENKINS_HOME
+    fi
     echo "Content in backup dir '${JENKINS_HOME_BACKUP_DIR}/${JENKINS_LAST_BACKUP}' copied to home dir '$JENKINS_HOME'"
   else
     echo "Backup found but existing data in jenkins home won't be overwritten"
